starting with rtc

+++ sw/src/util.c
@@ -4,7 +4,7 @@
     while (count--) asm("nop");
 }
 
-void int_to_str(uint32_t x, char *buf, uint8_t size, uint8_t radix) {
+void int_to_str(uint32_t x, char *buf, size_t size, uint8_t radix) {
     for (uint8_t i = size; i > 0; i--) {
         uint32_t rem = x % radix;
         x /= radix;
@@ -17,14 +17,14 @@ void int_to_str(uint32_t x, char *buf, uint8_t size, uint8_t radix) {
     }
 }
 
-bool str_cmp(char *buf1, char *buf2, uint8_t size) {
+bool str_cmp(char *buf1, char *buf2, size_t size) {
     for (uint8_t i = 0; i < size; i++) {
         if (buf1[i] != buf2[i]) return false;
     }
     return true;
 }
 
-void str_copy(char *buf1, char *buf2, uint8_t size) {
+void str_copy(char *buf1, char *buf2, size_t size) {
     for (uint8_t i = 0; i < size; i++) {
         buf2[i] = buf1[i];
     }
