+++ sw/src/util.h
@@ -16,10 +16,10 @@ typedef enum {
 
  __attribute__((optimize("O0"))) void spin(uint32_t count);
 
-void int_to_str(uint32_t x, char *buf, uint8_t size, uint8_t radix);
+void int_to_str(uint32_t x, char *buf, size_t size, uint8_t radix);
 
-bool str_cmp(char *buf1, char *buf2, uint8_t size);
+bool str_cmp(char *buf1, char *buf2, size_t size);
 
-void str_copy(char *buf1, char *buf2, uint8_t size);
+void str_copy(char *buf1, char *buf2, size_t size);
 
 #endif
